tests/test_main.py
"""Test suite for the RapidOCR FastAPI service."""

import io
from pathlib import Path
from unittest.mock import patch

import pytest
from fastapi.testclient import TestClient
from PIL import Image

from app.main import app

# Create test client
client = TestClient(app)


def create_test_image(
    format: str = "PNG", size: tuple[int, int] = (100, 50), color: str = "white"
) -> io.BytesIO:
    """Create a test image in memory."""
    image = Image.new("RGB", size, color)
    image_buffer = io.BytesIO()
    image.save(image_buffer, format=format)
    image_buffer.seek(0)
    return image_buffer


class TestHealthEndpoints:
    """Test health and status endpoints."""

    def test_root_endpoint(self) -> None:
        """Test the root endpoint returns basic info."""
        response = client.get("/")
        assert response.status_code == 200
        data = response.json()
        assert "service" in data
        assert "version" in data
        assert data["status"] == "running"

    def test_health_endpoint(self) -> None:
        """Test the health check endpoint."""
        response = client.get("/health")
        assert response.status_code == 200
        data = response.json()
        assert data["status"] == "healthy"
        assert "version" in data
        assert "gpu_available" in data
        assert "uptime" in data

    def test_stats_endpoint(self) -> None:
        """Test the stats endpoint."""
        response = client.get("/stats")
        assert response.status_code == 200
        data = response.json()
        assert "service" in data
        assert "gpu" in data
        assert "ocr_engine" in data
        assert "file_management" in data
        assert "configuration" in data


class TestOCREndpoints:
    """Test OCR processing endpoints."""

    def test_ocr_single_image(self) -> None:
        """Test OCR processing with a single image."""
        test_image = create_test_image()

        response = client.post(
            "/ocr", files={"files": ("test.png", test_image, "image/png")}
        )

        assert response.status_code == 200
        data = response.json()

        # Check response structure
        assert "results" in data
        assert "processing_time" in data
        assert "gpu_used" in data

        # Check results structure
        assert len(data["results"]) == 1
        result = data["results"][0]
        assert "FileName" in result
        assert "UUID" in result
        assert "Context" in result
        assert result["FileName"] == "test.png"

    def test_ocr_multiple_images(self) -> None:
        """Test OCR processing with multiple images."""
        test_image1 = create_test_image()
        test_image2 = create_test_image(color="red")

        response = client.post(
            "/ocr",
            files=[
                ("files", ("test1.png", test_image1, "image/png")),
                ("files", ("test2.png", test_image2, "image/png")),
            ],
        )

        assert response.status_code == 200
        data = response.json()

        # Check we got results for both images
        assert len(data["results"]) == 2

        # Check each result has the correct filename
        filenames = [result["FileName"] for result in data["results"]]
        assert "test1.png" in filenames
        assert "test2.png" in filenames

    def test_ocr_no_files(self) -> None:
        """Test OCR endpoint with no files."""
        response = client.post("/ocr")

        # Should return 422 (validation error) due to missing required files
        assert response.status_code == 422

    def test_ocr_too_many_files(self) -> None:
        """Test OCR endpoint with too many files."""
        # Create more files than the limit (default is 10)
        files = []
        for i in range(12):  # Exceed the default limit
            test_image = create_test_image()
            files.append(("files", (f"test{i}.png", test_image, "image/png")))

        response = client.post("/ocr", files=files)

        assert response.status_code == 400
        assert "Too many files" in response.json()["detail"]

    def test_ocr_large_file(self) -> None:
        """Test OCR endpoint with a file that's too large."""
        # Create a large image (this might not actually exceed the limit in memory)
        large_image = create_test_image(size=(2000, 2000))

        # We'll mock the file size check since creating an actually large file
        # in memory for testing is impractical
        with patch("app.main.settings.max_file_size", 1024):  # 1KB limit
            response = client.post(
                "/ocr", files={"files": ("large.png", large_image, "image/png")}
            )

            # May return 413 if the image is actually large enough
            # Otherwise, the test confirms the validation logic exists
            assert response.status_code in [200, 413]


class TestRequestLogging:
    """Test request logging and tracking."""

    def test_request_id_header(self) -> None:
        """Test that responses include a request ID header."""
        response = client.get("/health")
        assert response.status_code == 200
        assert "X-Request-ID" in response.headers

        # Request ID should be a valid UUID format
        request_id = response.headers["X-Request-ID"]
        assert len(request_id) == 36  # UUID string length
        assert request_id.count("-") == 4  # UUID format


class TestErrorHandling:
    """Test error handling and responses."""

    def test_invalid_endpoint(self) -> None:
        """Test accessing a non-existent endpoint."""
        response = client.get("/nonexistent")
        assert response.status_code == 404

    def test_invalid_method(self) -> None:
        """Test using wrong HTTP method."""
        response = client.put("/ocr")
        assert response.status_code == 405  # Method not allowed


@pytest.fixture
def test_image_path(tmp_path: Path) -> Path:
    """Create a test image file for testing."""
    image_path = tmp_path / "test_image.png"
    image = Image.new("RGB", (100, 50), "white")
    image.save(image_path)
    return image_path


class TestFileManagement:
    """Test file management functionality."""

    def test_temp_directory_creation(self, test_image_path: Path) -> None:
        """Test that temporary directories are properly managed."""
        # This is more of an integration test
        # The actual temp directory management is tested through the OCR endpoint

        with open(test_image_path, "rb") as f:
            response = client.post(
                "/ocr", files={"files": ("test.png", f, "image/png")}
            )

        assert response.status_code == 200
        # Files should be cleaned up automatically after processing
